+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Fix multi-dimensional parameter linking (bug introduced in 2.4.0 #594). #631
 - Fix bug where any argument containing an integer between commas would be interpreted as a frame range. #644
 - Python: `app.saveProject`  and `app.saveProjectAs` now do project variable substitution, as in `app.saveProjectAs("[Variable]/output.ntp")`.
+- Fix ASCII curve import. #656
 
 ### Plugins